refactor(common): move matomo-related logic back to browser-extension
