+++ packages/common/src/analytics/iAnalyticsTracker.ts
@@ -1,8 +1,7 @@
-import type { PostHogAction, MatomoSendEventProps, PostHogProperties } from './types';
+import type { PostHogAction, PostHogProperties } from './types';
 
 export interface IAnalyticsTracker {
   sendPageNavigationEvent: () => Promise<void>;
   sendAliasEvent: () => Promise<void>;
-  sendEventToMatomo: (props: MatomoSendEventProps) => Promise<void>;
   sendEventToPostHog: (action: PostHogAction, properties?: PostHogProperties) => Promise<void>;
 }
