+++ packages/common/src/analytics/types.ts
@@ -1,28 +1,3 @@
-/* eslint-disable camelcase */
-export enum MatomoEventActions {
-  CLICK_EVENT = 'click-event',
-  HOVER_EVENT = 'hover-event'
-}
-
-export enum MatomoEventCategories {
-  SEND_TRANSACTION = 'send-transaction',
-  WALLET_CREATE = 'wallet-create',
-  WALLET_RESTORE = 'wallet-restore',
-  HW_CONNECT = 'hw-connect',
-  VIEW_TOKENS = 'view-tokens',
-  VIEW_NFT = 'view-nft',
-  ADDRESS_BOOK = 'address-book',
-  VIEW_TRANSACTIONS = 'view-transactions',
-  STAKING = 'staking'
-}
-
-export type MatomoSendEventProps = {
-  category: MatomoEventCategories;
-  action: MatomoEventActions;
-  name: string;
-  value?: number;
-};
-
 export enum PostHogAction {
   // Hardware wallet connect
   OnboardingHWAnalyticsAgreeClick = 'onboarding | hardware wallet | analytics | agree | click',
